cinemapedia feat(app_router): bottom navigation bar implemented IndexedStack body: end of section 16B
