+++ cinemapedia/lib/presentation/widgets/movies/movie_horizontal_list_view.dart
@@ -122,7 +122,7 @@ class _Slide extends StatelessWidget {
                     );
                   }
                   return GestureDetector(
-                      onTap: () => context.push('/movie/${movie.id}'),
+                      onTap: () => context.push('/home/0/movie/${movie.id}'),
                       child: FadeIn(child: child));
                 },
               ),
