+++ cinemapedia/lib/presentation/widgets/shared/custom_app_bar.dart
@@ -37,7 +37,7 @@ class CustomAppBar extends ConsumerWidget {
                     ),
                   ).then((movie) {
                     if (movie?.id == null) return;
-                    context.push('/movie/${movie?.id}');
+                    context.push('/home/0/movie/${movie?.id}');
                   });
                 },
                 icon: const Icon(Icons.search)),
