+++ cinemapedia/lib/presentation/widgets/shared/custom_bottom_navigation_bar.dart
@@ -1,12 +1,21 @@
 import 'package:flutter/material.dart';
+import 'package:go_router/go_router.dart';
 
 class CustomBottomNavigationBar extends StatelessWidget {
-  const CustomBottomNavigationBar({super.key});
+  final int currentIndex;
+
+  const CustomBottomNavigationBar({super.key, required this.currentIndex});
+
+  void _onItemTapped(BuildContext context, int index) {
+    context.go('/home/$index');
+  }
 
   @override
   Widget build(BuildContext context) {
     return BottomNavigationBar(
       elevation: 0,
+      currentIndex: currentIndex,
+      onTap: (value) => _onItemTapped(context, value),
       items: const [
         BottomNavigationBarItem(
           label: 'Inicio',
